third_party/blink/web_tests/fast/forms/select/customizable-select/disallowed-select-descendants-console-message.html

<!DOCTYPE html>
<link rel=author href="mailto:ansollan@microsoft.com">
<link rel=help href="https://issues.chromium.org/issues/347890366">

<style>
  select, ::picker(select) {
    appearance: base-select;
  }
</style>

<!-- Should print two console messages. -->
<select>
  <label>label</label>
</select>

<!-- Should print two console messages. -->
<select>
  <input>
  <input>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <input>
  </option>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <datalist>
    </datalist>
  </option>
</select>

<!-- Should print one console messages. -->
<select>
  <option>
    <object>
    </object>
  </option>
</select>

<!-- Phrasing content but not Interactive content allowed. -->
<select>
  <option>
    <textarea></textarea>
  </option>
</select>

<!-- <button> (if present) should be the first child of <select>. -->
<select>
  <option></option>
  <option></option>
  <button>button</button>
</select>

<select>
  <option></option>
  <div>
    <button>button</button>
  </div>
  <span>
    <button>button</button>
  </span>
</select>

<!-- <legend> (if present) should be the first child of <optgroup>. -->
<select>
  <optgroup>
    <option>..</option>
    <legend>legend</legend>
  </optgroup>
</select>

<select>
  <optgroup>
    <div>
      <legend>legend</legend>
    </div>
  </optgroup>
  <optgroup>
    <span>
      <legend>legend</legend>
    </span>
  </optgroup>
</select>

<!-- <legend> elements should not have interactive elements inside -->
<select>
  <optgroup>
      <legend>
        <button></button>
      </legend>
  </optgroup>
</select>

<!-- <optgroup> should not have interactive elements inside. -->
<select>
  <optgroup>
    <option>..</option>
    <input/>
  </optgroup>
</select>

<!-- Should print two console messages. -->
<select>
  <div>
    <option>
      <span tabindex="1">..</span>
      <span tabindex="2">..</span>
    </option>
  </div>
</select>

<!-- Should print two console messages. -->
<select>
  <div>
    <option>
      <span contenteditable="true">..</span>
      <span contenteditable="plaintext-only">..</span>
    </option>
  </div>
</select>

<!-- Should print one console message. -->
<select class="delayed" style="display:none">
  <input>
  <option></option>
</select>

<!-- Should print two console messages. -->
<select class="delayed" style="display:none">
  <div>
    <option>
      <span tabindex="1">..</span>
      <span tabindex="2">..</span>
    </option>
  </div>
</select>

<!-- Should print two console messages. -->
<select class="delayed" style="display:none">
  <div>
    <option>
      <span contenteditable="true">..</span>
      <span contenteditable="plaintext-only">..</span>
    </option>
  </div>
</select>

<script>
  requestAnimationFrame(() => {
    document.querySelectorAll('.delayed').forEach((select) => {
      select.style.display = '';
    });
    requestAnimationFrame(() => {
        testRunner.notifyDone();
    });
  });
</script>

<select>
  <label is="custom-label"></label>
</select>

<script>
  class CustomLabel extends HTMLLabelElement {
    constructor() {
      super();
      this.style.color = 'blue';
      this.style.fontWeight = 'bold';
    }
  }
  customElements.define('custom-label', CustomLabel, { extends: 'label' });
</script>

<script>
testRunner.waitUntilDone();
testRunner.dumpAsText();
</script>
